feat: generateJsonDataFromMarkdownAtBuildTimeToImprovePerformance
Refactored the data loading mechanism. Instead of loading and parsing Markdown files at runtime using `import.meta.glob`, a new script (`scripts/generate-cheatsheet-data.ts`) parses all Markdown files in `src/data/go-cheatsheet-md/` and generates a single JSON file (`src/data/generated/cheatsheet-data.json`) during the build process (`npm run build` and `npm run dev`).

The frontend data loader (`src/data/markdown-cheatsheet-loader.ts`) now statically imports this generated JSON, significantly reducing the initial bundle size and eliminating runtime parsing overhead.

Added `glob` dependency for file searching in the generation script.

Updated `.gitignore` to exclude the generated data directory.

Closes: #37
