+++ .gitignore
@@ -26,3 +26,7 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+
+# Generated data
+src/data/generated/
